fix some warning issues

+++ BNCSQLite/Core/BNCSQLiteDatabase.h
@@ -93,7 +93,7 @@ extern NSInteger const kBNCSQLiteInitVersion;
 
  @return version of schema
  */
-- (NSInteger)currentVersion;
+- (UInt64)currentVersion;
 
 /**
  Update the schema version
